+++ tests/cmk_stubs/cmk/rulesets/v1/form_specs.py
@@ -5,14 +5,17 @@ class _Rec:
 
 
 Dictionary = DictElement = Integer = Float = String = BooleanChoice = _Rec
-SimpleLevels = LevelDirection = DefaultValue = Password = _Rec
+SimpleLevels = DefaultValue = Password = _Rec
 CascadingSingleChoice = CascadingSingleChoiceElement = _Rec
-SingleChoice = SingleChoiceElement = List = TimeSpan = TimeMagnitude = _Rec
+SingleChoice = SingleChoiceElement = List = TimeSpan = _Rec
 RegularExpression = validators = _Rec
 Percentage = _Rec
 
 MatchingScope = type("MatchingScope", (), {"PREFIX": 0, "INFIX": 1, "FULL": 2})
 LevelsType = type("LevelsType", (), {"NONE": 0, "FIXED": 1})
+LevelDirection = type("LevelDirection", (), {"UPPER": 0, "LOWER": 1})
+TimeMagnitude = type("TimeMagnitude", (), {
+    "MILLISECOND": 0, "SECOND": 1, "MINUTE": 2, "HOUR": 3, "DAY": 4})
 
 
 class _V:  # validators namespace